requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.55
botocore==1.40.55
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==3.24.0
databricks-sdk==0.68.0
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.119.0
fastjsonschema==2.21.2
fastmcp==2.12.5
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fredapi==0.5.2
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
missingno==0.5.2
mlflow==3.5.0
mlflow-skinny==3.5.0
mlflow-tracing==3.5.0
more-itertools==10.8.0
multitasking==0.0.12
narwhals==2.8.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.4
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
parse==1.20.2
parso==0.8.5
pathable==0.4.4
peewee==3.18.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.1
yfinance==0.2.66
zipp==3.23.0
